solved all bugs

+++ RDS-code/core.h
@@ -38,11 +38,11 @@ void prepare_quit();
 
 //db managing functions
 void append_to_csv(FILE* file, int argc, char* argv[]);
-char* search_bills(long CNP);
-int check_cnp(long CNP);
+char* search_bills(long long CNP);
+int check_cnp(long long CNP);
 void set_bill_data(char* bill_lines);
 char* get_full_csv_line(int number, char* filename);
-int does_client_exist(long cnp);
+int does_client_exist(long long cnp);
 
 //utility functions
 char* int_to_string(int x);
@@ -53,15 +53,15 @@ struct tm parse_date(char* in_string);
 int compare_dates(struct tm d1, struct tm d2, int level);
 void strapp(char* s, char c);
 void free_bill_data();
-int has_cable(long cnp);
+int has_cable(long long cnp);
 void map_phones();
-int has_phone(long cnp);
+int has_phone(long long cnp);
 inline void clear_console() { system("cls"); }
 char* get_config_property(char* key);
 void print_file_content(char* filename);
-void add_person(char* name, char* first_name, long cnp);
-void print_phone_contracts(long cnp);
-char** get_phone_numbers(long cnp);
+void add_person(char* name, char* first_name, long long cnp);
+void print_phone_contracts(long long cnp);
+char** get_phone_numbers(long long cnp);
 struct tm get_date();
 char* get_date_string(struct tm date);
