+++ RDS-code/objects.h
@@ -14,7 +14,7 @@ struct person
 	char name[41];
 	char first_name[41];
 	long long cnp;
-	void(*set_cnp)(person*, long);
+	void(*set_cnp)(person*, long long);
 	void(*set_first_name)(person*, char*);
 	void(*set_name)(person*, char*);
 	void(*print)(person*);
@@ -42,7 +42,7 @@ void set_person_name(person* pers, char* name);
 void set_person_first_name(person* pers, char* first_name);
 void set_person_cnp(person* pers, long long cnp);
 void print_person(person* pers);
-bill* bill_init(char* name, char* first_name, char* address, service_type service, long cnp, struct tm date, int sum);
+bill* bill_init(char* name, char* first_name, char* address, service_type service, long long cnp, struct tm date, int sum);
 
 void set_bill_name(bill* obj, char* name);
 void set_bill_first_name(bill* obj, char* first_name);
